refactor: modified database connection configuration
